Raffaele Sena: As promised, here is the HTML parser updated to support all the tags and attributes defined in HTML 4.0.
I left all the deprecated tags, but removed all the tags that
basically never existed in the real world (H7 ? HTMLPLUS, etc.)

I added all the attributes as specified in the HTML 4.0 DTD, apart for
the SCRIPT/event related attributes ( ON* ).

I created a couple of macros to define attributes, so that basically
the same definition can be copied between HTMLPDTD.c and HTMLPDTD.h
(preserving the ordering and the spelling).